refactor: convert MyChildrenList and ChildCard to Server Components
